Restructure application to fit proper Flask layout
