+++ clubWebsite/templates/faq.html
