+++ wsgi.py
@@ -1,4 +1,5 @@
 from clubWebsite import app
+from clubWebsite import routes
 
 if __name__ == '__main__':
     app.run('localhost', port=8080, debug=True)
